Replaces DataModel class with PipelineResult
