+++ requirements.txt
@@ -1,5 +1,6 @@
 astropy>=4.0
 bokeh
+egon
 matplotlib
 iminuit<2.0.0
 joblib
