+++ tests/test_nodes.py
@@ -1,9 +1,10 @@
+from tempfile import NamedTemporaryFile
 from unittest import TestCase
 
 import numpy as np
 from egon.mock import MockTarget
 
-from snat_sim.pipeline import LoadPlasticcSims, plasticc
+from snat_sim.pipeline import LoadPlasticcSims, plasticc, FitResultsToDisk, DataModel, SNModel
 
 
 class TestLoadPlasticcSims(TestCase):
@@ -23,3 +24,15 @@ class TestLoadPlasticcSims(TestCase):
         plasticc_data = next(plasticc.iter_lc_for_cadence_model(self.cadence, model=11))
         loaded_data = next(self.mock_target.input.iter_get())
         np.testing.assert_array_equal(plasticc_data, loaded_data)
+
+
+class TestFitResultsToDisk(TestCase):
+
+    def setUp(self) -> None:
+        sn_model = SNModel('salt2')
+        self.out_file = NamedTemporaryFile()
+        self.data_model = DataModel(sn_model, sn_model)
+        self.write_action = FitResultsToDisk(self.data_model, self.out_file.name)
+
+    def tearDown(self) -> None:
+        self.out_file.close()
